requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.15.0
